edited files
added 2 more functions (ft_strnstr.c, ft_memcmp.c)
edited the Makefile
edited the libft.h

+++ ft_strlcat.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char *dst, char *src, unsigned int size)
+size_t	ft_strlcat(char *dst, char *src, unsigned int size)
 {
 	unsigned int	i;
 	unsigned int	j;
